+++ react/components/Footer/propTypes.js
@@ -1,15 +1,15 @@
 import PropTypes from 'prop-types'
 
-export const objectLikeLinkArray = PropTypes.objectOf(PropTypes.shape({
+export const objectLikeLinkArray = PropTypes.arrayOf(PropTypes.shape({
   url: PropTypes.string,
   title: PropTypes.string,
 }))
 
-export const objectLikeBadgeArray = PropTypes.objectOf(PropTypes.shape({
+export const objectLikeBadgeArray = PropTypes.arrayOf(PropTypes.shape({
   image: PropTypes.string,
 }))
 
-export const objectLikePaymentFormArray = PropTypes.objectOf(PropTypes.shape({
+export const objectLikePaymentFormArray = PropTypes.arrayOf(PropTypes.shape({
   paymentType: PropTypes.oneOf([
     'MasterCard', 'Visa', 'Diners Club',
   ]),