+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Documentation of `BuyButton`.
 
+### Changed
+- Migrate the `Footer` schema to use the `type: "array"`.
+
 ## [1.8.0] - 2018-07-10
 ### Added
 - Buy button support to add multiple sku items to the cart at the same time of `BuyButton`.